Experimental language parsing
This is a test implementation of the brainstorm around parsing of languages at the Microformats wiki: http://microformats.org/wiki/microformats2-parsing-brainstorming#Parse_language_information

The tests are added here for now even though I know that they belong in a separate repository eventually if this brainstorm reaches a conclusion. For now its easier to keep it all in the same place.